Change the way FileEntry and restores work:
internally, files to backup are stored without path. if there are multiple files
with the same file name, append an increasing number to those file names
Core.BackupFiles is now an HashSet instead of Dictionary. Key is internal file name
Original file name (with path) is stored in FileEntry: it will be useful when disambiguating
the file to restore (user inputs the filename without path)
also, restored files will be put in a 'restore' directory (configurable in config)